PRStats/Utils.cs
using System.Net.Http;
using System.Net.Http.Headers;

namespace PRStats
{
    public static class Utils
    {
        public static HttpClient getGHClient(string token)
        {
            var client = new HttpClient();
            client.DefaultRequestHeaders.Accept.Clear();
            client.DefaultRequestHeaders.Accept.Add(
                new MediaTypeWithQualityHeaderValue("application/vnd.github.v3+json"));
            client.DefaultRequestHeaders.Add("User-Agent", "PRStats-Client");
            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
            return client;
        }
    }
}